requirements.txt
black==23.3.0
pycryptodome==3.18.0